requirements.txt
# Core
crewai>=0.80.0
crewai-tools>=0.15.0
anthropic>=0.40.0
langchain-anthropic>=0.3.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Telegram
python-telegram-bot>=21.0.0

# Supabase
supabase>=2.3.0

# PDF Generation
reportlab>=4.2.0
Pillow>=10.4.0

# HTTP
httpx>=0.27.0
requests>=2.31.0
tenacity>=8.3.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
aiofiles>=23.2.0
python-dateutil>=2.9.0
pytz>=2024.1
